requirements.txt
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn==0.27.0
playwright==1.41.0
apscheduler==3.10.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
aiohttp==3.9.1
pillow==10.2.0
openai==1.10.0
